services/web-ui/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
jinja2==3.1.4
kafka-python==2.0.2